api/requirements.txt
flask
werkzeug
PyJWT
flask-sqlalchemy
sqlalchemy
marshmallow